+++ root_orchestrator/resource-abstractor/api/v1/apps_blueprint.py
@@ -1,10 +1,11 @@
 from bson import json_util
-from db import apps_db
+from db import jobs_db as apps_db
 from flask import request
 from flask.views import MethodView
 from flask_restful import Resource
 from flask_smorest import Blueprint
 from marshmallow import INCLUDE, Schema, fields
+from services import mediator
 
 applicationblp = Blueprint(
     "Application operations",
@@ -40,7 +41,7 @@ class CreateApplicationController(Resource):
     )
     def post(self, userid, *args, **kwargs):
         data = request.get_json()
-        return apps_db.create_app(userid, data)
+        return mediator.perform_create(apps_db.create_app, data, userid, entity="application")
 
 
 @applicationblp.route("/<userid>/<appid>")
@@ -53,18 +54,15 @@ class ApplicationController(MethodView):
 
     @applicationblp.response(200, content_type="application/json")
     def delete(self, userid, appid, *args, **kwargs):
-        return json_util.dumps(apps_db.delete_app(userid, appid))
-
-    @applicationblp.response(
-        200, ApplicationSchema(unknown=INCLUDE), content_type="application/json"
-    )
-    def put(self, userid, appid, *args, **kwargs):
-        data = request.get_json()
-        return apps_db.update_app(userid, appid, data)
+        return json_util.dumps(
+            mediator.perform_delete(apps_db.delete_app, appid, userid, entity="application")
+        )
 
     @applicationblp.response(
         200, ApplicationSchema(unknown=INCLUDE), content_type="application/json"
     )
     def patch(self, userid, appid, *args, **kwargs):
         data = request.get_json()
-        return apps_db.update_app(userid, appid, data)
+        return mediator.perform_update(
+            apps_db.update_app, appid, data, userid, entity="application"
+        )
